+++ packages/mongo_document_annotation/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.4.9
+
+### Fixed
+Bug fixes around Base Projections
+
 ## 1.4.8
 
 ### Added